#32 Fixed class names
